Move dce to analysis
